correciones

+++ src/main/kotlin/Punto.kt
@@ -1,7 +1,7 @@
-data class Punto(var x: Int, var y: Int) {
+data class Punto(var x: Double, var y: Double) {
 
 }
 
-data class ParDePuntos(val p1: Punto, val p2: Punto, val d: Double){
+data class ParDePuntos(val p1: Punto, val p2: Punto, val d: Double ){
 
 }
